refactor(Main) : #13 - 폴더 구조 변경
메인 페이지에 있는 컴포넌트를 feature 폴더로 분리

Related to: #13
