+++ frontend/src/common/constants/Pages.ts
@@ -1,0 +1,7 @@
+const Pages = {
+    LOGIN: 'login',
+    SIGN_UP: 'signup',
+    MAIN: 'main',
+} as const;
+
+export default Pages;
