+++ frontend/src/features/Main/components/Carousel/Carousel.tsx
